bring ozones within the template format
